requirements.txt
alembic==1.13.1
asgiref==3.8.1
cachetools==4.2.4
click==8.1.7
ecdsa==0.19.0
fastapi==0.68.0
greenlet==3.0.3
h11==0.14.0
Mako==1.3.3
MarkupSafe==2.1.5
passlib==1.7.4
pyasn1==0.6.0
pydantic==1.9.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.25
starlette==0.14.2
typing_extensions==4.11.0
uvicorn==0.15.0
